requirements.txt
Jinja2
PyYaml